Pull out AppMode.
